refactor: removes weird non-lua syntax
reworks from old lovepack standards (which sucked) to a more lua way to
do a library. now the whole folder is considered the library, and you
can just move the folder where you want it, and require it and it should
work.

had to update the test to not looked for a compiled library, and instead
look for this folder in the root of this folder?!?!

also got rid of the weird lib.toml file i was trying to do. probably not
the best way to do the library thing because it would alienate people
from using the library because they would have to use my library
management system (and they probably wouldn't want to).